Add unit tests for base input class
